requirements.txt
Pillow==9.3.0
python-barcode==0.14.0